Move MetadataSyncUtils methods back to DataFeeds
